add tests to existing file
